Edit to title tags + Home page
I edited the home page to get rid of a piece of javascript which I had
included to satisfy a requirement for a project in one of my classes.
This piece of javascript was not on the my website online (as I had
already removed it) but it was still in the files on my computer.

I also edited the titles all of the pages to add 'Osman Sufi - ' to the
beginning of each page title. So, for example, the home page should now
say 'Osman Sufi - Home'