Redesign for "Mobile First" initiated
- Altered CSS file so that the website changes orientation depending on
whether it is being viewed on a phone, computer, or tablet
- Problem with container not extending to the bottom of the page unless
I div tag with the footer id is included. This came about as a result of
the changes made in this version. It has something to do with the fact
that the left and right ids contain the float attribute (with the value
"left"). If the float attribute is gotten rid of on either it 'extends'
the container over them (although, in a very ugly way). This, however,
messes up the formatting. But why is it doing this? Until a real fix is
found the div tag with footer id will have to do...
- Corrected an error in an in-page link for the blog page
- Fixed gif size on hobbies page for phone screen orientation